rag/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
pypdf>=4.2.0
